develops_today_test_project/requirements.txt
#asgiref==3.3.1
#Django==3.1.4
djangorestframework==3.12.2
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1